Get rid of the term projection.
Pierre pointed out that "chunks" or "fragments" might be
better terminology. Jean-Pierre voted for "chunk".

I was going to go with "chunk", but I liked the operation
WRITE_BLOCK and not WRITE_CHUNK. I think "block" flows
better in the reading as well. As such, we have
simply header, block, and data block.

Signed-off-by: Tom Haynes <loghyr@gmail.com>