Rewrite the Introduction
Christoph raised the point that the whilst the approach
was sound, the implementation was not the best.

Tom (myself) was also unhappy with the complexity
being generated with the new ops.

As such, the draft is being revamped to make
use of atomic swapping of byte ranges in two
files.

Signed-off-by: Tom Haynes <loghyr@hammerspace.com>